Remove quotes from Regexp when printing error messages.
Already found a way to make this look nicer. Since the `@css_selector` is purely for
presentation and never hits Nokogiri it doesn't have to be in quotes and we can make it
look like the what the user put in.

Replaces `div:match\('id',\ "\/wups\/"\)` with `div:match('id', /wups/)`.