Raise an error if the last arg is the wrong format
I always accidentally write this:

```ruby
assert_select "title", "Welcome", count: 1
```

When I mean to write this:

```ruby
assert_select "title", text: "Welcome", count: 1
```

The first assertion is incorrect; it's looking for the string "Welcome", then using `{ count: 1}` as the assertion failure message if it's not found.

Since the assertion failure message probably should not be a hash, I propose we raise an error if that argument is passed. This would make it easier to identify incorrectly written tests.